+++ vertexai/app/src/main/kotlin/com/formulae/chef/feature/chat/ui/ChefOverlay.kt
@@ -47,6 +47,8 @@ import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.unit.dp
 import com.formulae.chef.feature.chat.OverlayChatViewModel
 import com.formulae.chef.feature.model.Recipe
+import com.formulae.chef.services.voice.TTS_DISPLAY_THRESHOLD
+import com.formulae.chef.services.voice.sanitizeMarkdown
 import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterial3Api::class)
