+++ vertexai/app/src/main/kotlin/com/formulae/chef/feature/chat/ui/VoiceController.kt
@@ -23,6 +23,8 @@ import com.formulae.chef.BuildConfig
 import com.formulae.chef.services.voice.AudioPlayer
 import com.formulae.chef.services.voice.GcpTextToSpeechService
 import com.formulae.chef.services.voice.SpeechInputManager
+import com.formulae.chef.services.voice.TTS_DISPLAY_THRESHOLD
+import com.formulae.chef.services.voice.sanitizeForTts
 import kotlinx.coroutines.launch
 
 data class VoiceControllerState(
