+++ vertexai/app/src/main/kotlin/com/formulae/chef/services/voice/TextSanitizer.kt
@@ -1,4 +1,4 @@
-package com.formulae.chef.feature.chat.ui
+package com.formulae.chef.services.voice
 
 private val MARKDOWN_CHARS = Regex("[*#`_~]")
